requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary>=2.9.0
alembic==1.13.1
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
# uuid is part of Python standard library